+++ tests/integration/__init__.py
@@ -18,6 +18,7 @@ except ImportError as e:
 
 CLUSTER_NAME = 'test_cluster'
 CCM_CLUSTER = None
+DEFAULT_CASSANDRA_VERSION = '1.2.9'
 
 path = os.path.join(os.path.abspath(os.path.dirname(__file__)), 'ccm')
 if not os.path.exists(path):
@@ -58,14 +59,16 @@ def get_node(node_id):
 
 
 def setup_package():
+    version = os.getenv("CASSANDRA_VERSION", DEFAULT_CASSANDRA_VERSION)
     try:
         try:
             cluster = CCMCluster.load(path, CLUSTER_NAME)
             log.debug("Found existing ccm test cluster, clearing")
             cluster.clear()
+            cluster.set_cassandra_dir(cassandra_version=version)
         except Exception:
-            log.debug("Creating new ccm test cluster")
-            cluster = CCMCluster(path, CLUSTER_NAME, cassandra_version='1.2.9')
+            log.debug("Creating new ccm test cluster with version %s", version)
+            cluster = CCMCluster(path, CLUSTER_NAME, cassandra_version=version)
             cluster.set_configuration_options({'start_native_transport': True})
             common.switch_cluster(path, CLUSTER_NAME)
             cluster.populate(3)